requirements.txt
Flask
flask-cors
opencv-python
mediapipe
tensorflow-cpu
numpy
gunicorn
